fix(pipeline): cast numeric scores in matcher update and seed workspace isolation embeddings
Add explicit numeric casting to deterministic match score updates in PostgreSQL, and provision complete workspace embedding fixtures for isolation test.

+++ src/pipeline/deterministicMatcher.ts
@@ -1000,8 +1000,8 @@ export async function runDeterministicMatcher(
 
         const canonicalUpdate = await client.query(
           `UPDATE canonical_jobs
-           SET deterministic_match_score = CASE WHEN $8::boolean THEN NULL ELSE $2 END,
-               deterministic_match_coverage = CASE WHEN $8::boolean THEN NULL ELSE $3 END,
+           SET deterministic_match_score = CASE WHEN $8::boolean THEN NULL ELSE $2::numeric END,
+               deterministic_match_coverage = CASE WHEN $8::boolean THEN NULL ELSE $3::numeric END,
                latest_match_run_id = $4,
                profile_match_status = $6,
                processing_state = CASE WHEN $8::boolean THEN 'LANE_ROUTED' ELSE 'MATCHED' END,
